Upload `_3DProduct` with specified status.
- Enable providing the status in `_Sumbit.json` of `RFProduct._3D`.
- Implement support for uploading `_3DProduct` as draft.
- Don't upload `RFProduct._3D` if its status is `none`.
- Write status from `_Sumbit.json` to `_Status.json`.
- Store draft ID instead of product ID in `turbosquid.meta` for drafts.

+++ _3DProductsPublish/_3DProductDS/_3DProduct.cs
@@ -62,7 +62,7 @@ public partial record _3DProduct : IDisposable
 
 public record _3DProduct<TMetadata> : _3DProduct
 {
-    public int ID { get; internal set; } = default;
+    public long ID { get; internal set; } = default;
     internal _3DProduct(_3DProduct _3DProduct, TMetadata metadata)
         : base(_3DProduct)
     { Metadata = metadata; }
